🏗️ Tore the `string` measurement type apart.
