+++ test/unit/string/unit.test.ts
@@ -1,0 +1,8 @@
+import {Measurement, unit} from "!src/string";
+
+describe('unit', () => {
+  it("should return the unit of the measurement.", () => {
+    const measurement: Measurement<"px", 12.5> = "12.5px" as const;
+    expect(unit(measurement)).toBe("px");
+  });
+});
